Create index.md
The motivation behind this meme is how I am currently approaching an assignment for my 220 Statistics course. I used inspiration from existing memes but adapted text, images and colour to my situation. The code for the meme is an adaption of an existing code for a meme however, there is original code used. An example of this can be seen in line 22 where I have assigned a name to an objects; this allowed me to reuse the code.